Improves spec organization
The goal here is to have all describe blocks look like method names.
They should not include what really belongs in a context block.

Then, adds a second spec to verify that the `nil` json definition is
being used in the inner context, rather than the real json definition in
the outer context.